organise Gltf CatlogItem
