+++ lib/Models/Catalog/Gltf/GLTF.ts
