+++ lib/Models/Catalog/Gltf/GltfCatalogItem.ts
