+++ test/Models/Catalog/CatalogItems/GltfCatalogItemSpec.ts
@@ -3,7 +3,7 @@ import CesiumMath from "terriajs-cesium/Source/Core/Math";
 import Transforms from "terriajs-cesium/Source/Core/Transforms";
 import Entity from "terriajs-cesium/Source/DataSources/Entity";
 import MappableMixin from "../../../../lib/ModelMixins/MappableMixin";
-import GltfCatalogItem from "../../../../lib/Models/Catalog/CatalogItems/GltfCatalogItem";
+import GltfCatalogItem from "../../../../lib/Models/Catalog/Gltf/GltfCatalogItem";
 import CommonStrata from "../../../../lib/Models/Definition/CommonStrata";
 import createStratumInstance from "../../../../lib/Models/Definition/createStratumInstance";
 import Terria from "../../../../lib/Models/Terria";
